+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All Notable changes to `Manager` will be documented in this file
  - Cleanup and refactor
  - add `has('$dep.whatever')` interpolation
  - add linking (multiple aliases) to ioc manager
+ - add to ioc manager, fetch a class by string
  
 # v0.8.9 - 3-11-2016
  - Add full composer.json